Switch market sync.Map to regular map
The usage of sync.Map was premature. I don't think there is a use case to update the map itself while running (like adding a new stock item), so no need to have locking on it outside of the regular market lock for the data it contains. If there ever is a need to add data live, I can either use the existing market lock or add the sync.Map back in.

I plan to do the same thing for the other sync.Map types.